Extrapolate runtime-common mock runtime
