+++ runtimes/common/src/dip/did/tests.rs
@@ -22,9 +22,12 @@ use pallet_dip_provider::traits::IdentityProvider;
 
 use crate::{
 	constants::dip_provider::MAX_LINKED_ACCOUNTS,
-	dip::did::{
-		mock::{create_linked_info, ExtBuilder, TestRuntime, ACCOUNT, DID_IDENTIFIER, SUBMITTER},
-		LinkedDidInfoOf, LinkedDidInfoProvider, LinkedDidInfoProviderError, Web3OwnershipOf,
+	dip::{
+		did::{
+			mock::{create_linked_info, ExtBuilder, ACCOUNT, DID_IDENTIFIER, SUBMITTER},
+			LinkedDidInfoOf, LinkedDidInfoProvider, LinkedDidInfoProviderError, Web3OwnershipOf,
+		},
+		mock::TestRuntime,
 	},
 };
